test: Add unit test for checking the validity of student names
