--- images/api/__test__/unit/helpers.test.js
@@ -1,13 +1,0 @@
-const {
-  checkStudentPostEndpoint,
-} = require("../../src/helpers/endpointHelpers");
-
-/**table.uuid("UUID");
-    table.string("name");
-    table.integer("age");
-    table.string("classgroup"); // Class group information
-    table.double("grade"); // Student's grade information
-
-*/
-
-test("check name", () => {});
