+++ images/api/package.json
@@ -2,11 +2,11 @@
   "name": "portfolio-starter-amauryrw",
   "version": "1.0.0",
   "description": "",
-  "main": "src/routes/index.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "nodemon src/routes/index.js",
     "test": "jest --watchAll",
-    "test:integration": "jest __tests__/integration.test.js"
+    "start": " npx knex migrate:latest --cwd ./src/db && nodemon src/index.js",
+    "dev": " npx knex migrate:latest --cwd ./src/db && npx knex seed:run --cwd ./src/db && nodemon src/index.js"
   },
   "author": "",
   "license": "ISC",
@@ -16,9 +16,12 @@
     "knex": "^3.0.1",
     "mongodb": "^6.1.0",
     "mysql": "^2.18.1",
-    "nodemon": "^3.0.1"
+    "mysql2": "^3.6.5",
+    "nodemon": "^3.0.1",
+    "pg": "^8.11.3"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "supertest": "^6.3.3"
   }
 }
