+++ images/api/src/app.js
@@ -2,10 +2,10 @@
 const express = require('express'); 
 const knex = require('knex');
 const bodyParser = require('body-parser');
-const initEndpoints = require("./index.js");
+const initEndpoints = require("./routes");
 
 // require and manage database connection
-const knexfile = require('./db/knexfile'); // Import your Knex configura const db
+const knexfile = require('./db/knexfile.js'); // Import your Knex configura const db
 const db = knex(knexfile.development);
 
 // initialise app and routes; const app = express();
