+++ images/api/src/db/knexfile.js
@@ -1,0 +1,24 @@
+// Update with your config settings.
+
+/**
+ * @type { Object.<string, import("knex").Knex.Config> }
+ */
+module.exports = {
+  development: {
+    client: "pg",
+    connection: {
+      host: "localhost",
+      user: "your_username",
+      password: "your_password",
+      database: "your_database",
+    },
+    migrations: {
+      tableName: "knex_migrations",
+      directory: "./migrations",
+    },
+    seeds: {
+      directory: "./seeds",
+    },
+  },
+
+};
