+++ images/api/src/helpers/endpointHelpers.js
@@ -18,6 +18,9 @@ function checkStudentName(name) {
   return true;
 }
 
+
+
 module.exports = {
   checkStudentName,
+ 
 };
